refactor: 기물 패키지 분류
