+++ src/main/java/chess/domain/piece/directionmove/Bishop.java
@@ -1,6 +1,8 @@
-package chess.domain.piece;
+package chess.domain.piece.directionmove;
 
 import chess.domain.board.Direction;
+import chess.domain.piece.PieceType;
+import chess.domain.piece.Team;
 import java.util.Set;
 
 public class Bishop extends DirectionMovePiece {
