+++ src/main/java/chess/domain/piece/pawn/NormalBlackPawn.java
@@ -1,5 +1,6 @@
-package chess.domain.piece;
+package chess.domain.piece.pawn;
 
+import chess.domain.piece.Team;
 import java.util.Map;
 import java.util.Map.Entry;
 import java.util.Set;
